requirements.txt
gradio
requests
python-dotenv
langchain
langchain-google-genai
langchain-core
langchain-community
pandas
langgraph
langchain-anthropic
pypdf
youtube-transcript-api
wikipedia
duckduckgo-search
transformers
langchain-experimental
unstructured
openpyxl
assemblyai
langchain-groq
torch
yt-dlp
beautifulsoup4
google-genai